+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.9.19
+
+* fixed typing in `import_from`
+