+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # v2.9.19
 
-* fixed typing in `import_from`
+* added `length`