+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # v2.9.20
 
-
-
+* fixed typing annotations in `satella.coding.sequences`
+* added `merge_series`